icompatiblityConstraint and BackTracking has been modified
changed filter method of IcompatibilityConstraint and filterDomain of
BackTracking to fit the text.

BackTracking's method filterDomain need to be change. When the size of
the domain is 1 filterdomain has to put the value of the variable in the
car. he do it but don't give to us enough value (3 in place of 16)